Casting issue resolved

+++ service-loans/src/main/java/com/capitaworld/service/loans/repository/common/LoanRepository.java
@@ -110,5 +110,5 @@ public interface LoanRepository{
     
     public Boolean isNoBankStatement(Long applicationId);
     
-    public Double getMinRelationshipInMonthByApplicationId(Long applicationId);
+    public Integer getMinRelationshipInMonthByApplicationId(Long applicationId);
 }
